Removed autoscroll, added scrollbar, edited styling
small change on what we do slides

added scrollbar for aesthetics

toggled off autoscroll which is the flicking scroll on a smartphone and also moves the website to a new subheader when you move the scroll wheel.

now it should behave better on phones for sure as before you could not see all the text on a subheader of the site.